backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-JWT-Extended
psycopg2-binary
python-dotenv
bcrypt
